Full Lasers
Fully implemented the laser system. each laser object has a line object.
the line simply contains the start and end point coordinates. when a
laser is created it is added to an array that the render loop iterates
through to render all lasers in that array. Also created a brush object
that extends the graphics object, this is so I can easily draw lines by
passing in my own line object rather than having to use the default
drawLine method, that only takes in points to start and stop the line.